+++ src/main/java/kitchenpos/repository/MenuRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain.repository;
+package kitchenpos.repository;
 
 import kitchenpos.domain.Menu;
 import org.springframework.data.jpa.repository.JpaRepository;
