+++ src/main/java/kitchenpos/repository/OrderRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain.repository;
+package kitchenpos.repository;
 
 import kitchenpos.domain.Order;
 import kitchenpos.domain.OrderStatus;
