+++ src/main/java/kitchenpos/repository/ProductRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain.repository;
+package kitchenpos.repository;
 
 import kitchenpos.domain.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
