+++ src/main/java/kitchenpos/repository/TableGroupRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain.repository;
+package kitchenpos.repository;
 
 import kitchenpos.domain.TableGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
