+++ src/test/java/kitchenpos/application/MenuServiceTest.java
@@ -4,9 +4,9 @@ import kitchenpos.domain.Menu;
 import kitchenpos.domain.MenuGroup;
 import kitchenpos.domain.MenuProduct;
 import kitchenpos.domain.Product;
-import kitchenpos.domain.repository.MenuGroupRepository;
-import kitchenpos.domain.repository.MenuRepository;
-import kitchenpos.domain.repository.ProductRepository;
+import kitchenpos.repository.MenuGroupRepository;
+import kitchenpos.repository.MenuRepository;
+import kitchenpos.repository.ProductRepository;
 import kitchenpos.domain.MenuProducts;
 import kitchenpos.ui.dto.CreateMenuProductRequest;
 import kitchenpos.ui.dto.CreateMenuRequest;
